# Is StreamView::freeSectors needed?