--- a/sources/OpenMcdf/StreamView.cs
+++ b/sources/OpenMcdf/StreamView.cs
@@ -22,8 +22,6 @@ namespace OpenMcdf
         private long position;
         private readonly Stream stream;
         private readonly bool isFatStream;
-        private readonly List<Sector> freeSectors = new();
-        public IEnumerable<Sector> FreeSectors => freeSectors;
 
         public StreamView(List<Sector> sectorChain, int sectorSize, Stream stream)
         {