@Numpsy 2024-10-30
The ```InMemory``` benchmark here only goes from 
```
| Method | BufferSize | TotalStreamSize | Mean        | Error     | StdDev      | Median      | Gen0      | Gen1     | Gen2     | Allocated |
|------- |----------- |---------------- |------------:|----------:|------------:|------------:|----------:|---------:|---------:|----------:|
| Test   | 1048576    | 1048576         |    158.7 us |   3.12 us |     4.77 us |    158.1 us |  181.1523 |  90.5762 |        - |   1.09 MB |
| Test   | 524288     | 1048576         |    168.5 us |   2.68 us |     2.37 us |    168.1 us |  184.0820 |  92.0410 |        - |    1.1 MB |
| Test   | 262144     | 1048576         |    181.7 us |   3.41 us |     2.85 us |    180.7 us |  189.6973 |  88.1348 |        - |   1.14 MB |
| Test   | 131072     | 1048576         |    228.6 us |   3.97 us |     3.52 us |    227.4 us |  200.9277 | 100.3418 |        - |    1.2 MB |
| Test   | 4096       | 1048576         |  2,923.7 us |  47.46 us |    66.53 us |  2,907.0 us |  898.4375 | 398.4375 |        - |   5.39 MB |
| Test   | 1024       | 1048576         | 11,271.1 us | 208.83 us |   185.12 us | 11,246.9 us | 3062.5000 | 343.7500 | 171.8750 |  18.33 MB |
| Test   | 512        | 1048576         | 24,782.9 us | 596.95 us | 1,693.45 us | 24,280.3 us | 5937.5000 | 343.7500 | 156.2500 |   35.6 MB |
```

to 
```
| Method | BufferSize | TotalStreamSize | Mean        | Error     | StdDev    | Median      | Gen0      | Gen1     | Gen2     | Allocated |
|------- |----------- |---------------- |------------:|----------:|----------:|------------:|----------:|---------:|---------:|----------:|
| Test   | 524288     | 1048576         |    171.0 us |   3.38 us |   7.71 us |    168.7 us |  184.0820 |  92.0410 |        - |    1.1 MB |
| Test   | 1048576    | 1048576         |    180.5 us |   9.76 us |  28.79 us |    188.8 us |  181.1523 |  90.5762 |        - |   1.09 MB |
| Test   | 262144     | 1048576         |    183.1 us |   3.61 us |   3.38 us |    183.4 us |  189.6973 |  87.1582 |        - |   1.14 MB |
| Test   | 131072     | 1048576         |    230.8 us |   4.57 us |   9.24 us |    227.9 us |  200.9277 | 100.3418 |        - |    1.2 MB |
| Test   | 4096       | 1048576         |  2,961.0 us |  58.71 us | 131.31 us |  2,910.7 us |  894.5313 | 378.9063 |        - |   5.37 MB |
| Test   | 1024       | 1048576         | 11,222.6 us | 219.51 us | 328.55 us | 11,231.8 us | 3046.8750 | 359.3750 | 171.8750 |  18.26 MB |
| Test   | 512        | 1048576         | 23,522.0 us | 461.55 us | 631.78 us | 23,304.1 us | 5906.2500 | 406.2500 | 187.5000 |  35.44 MB |
```
for another test