@jeremy-visionaid 2024-11-14
@Numpsy 3.0.0 is pretty close now, we could probably just close this one unless it's a problem for you to switch to the new API?