@Numpsy 2024-11-14
Was just an observation, I haven;t had the time to test the v3 test yet