@Numpsy 2024-11-14
> unless it's a problem for you to switch to the new API

Personally I'd find it useful to have a new official build to get https://github.com/ironfede/openmcdf/pull/128 for maintenance purposes and take v3 after I do a major version bump and drop .NET 6.0 support.
This could go into such a build, but it's only a small allocation improvement so it's not a problem to not take the change