@jeremy-visionaid 2024-11-14
@Numpsy, I think we should get 2.4 out of the door too. @ironfede has granted me NuGet privileges, so I'll try and get that done now... I might leave this one here till after 3.0 is out, since it's probably easier for me to retarget this to a 2.4 branch.